+++ Solutions/Digital Shadows/Data Connectors/DigitalShadowsConnectorAzureFunction/AS_api.py
