+++ Solutions/Digital Shadows/Data Connectors/DigitalShadowsConnectorAzureFunction/AS_poller.py
