+++ Solutions/Digital Shadows/Data Connectors/DigitalShadowsConnectorAzureFunction/constant.py
