+++ Solutions/Digital Shadows/Data Connectors/DigitalShadowsConnectorAzureFunction/function.json
