+++ Solutions/Digital Shadows/Data Connectors/DigitalShadowsConnectorAzureFunction/state_serializer.py
