+++ Solutions/Digital Shadows/Data Connectors/DigitalShadowsSearchlight_API_functionApp.json
@@ -119,7 +119,7 @@
         },
         {
             "title": "2. Import Function App Code(Zip deployment)",
-            "description": "1. Install Azure CLI\n2. From terminal type **az functionapp deployment source config-zip -g <ResourceGroup> -n <FunctionApp> --src <Zip File>** and hit enter. Set the `ResourceGroup` value to: your resource group name. Set the `FunctionApp` value to: your newly created function app name. Set the `Zip File` value to: `digitalshadowsConnector.zip`(path to your zip file). Note:- Download the zip file from the link - [Function App Code](https://github.com/Azure/Azure-Sentinel/blob/master/Solutions/Digital%20Shadows/Data%20Connectors/Digital%20Shadows/digitalshadowsConnector.zip)"
+            "description": "1. Install Azure CLI\n2. From terminal type **az functionapp deployment source config-zip -g <ResourceGroup> -n <FunctionApp> --src <Zip File>** and hit enter. Set the `ResourceGroup` value to: your resource group name. Set the `FunctionApp` value to: your newly created function app name. Set the `Zip File` value to: `digitalshadowsConnector.zip`(path to your zip file). Note:- Download the zip file from the link - [Function App Code](https://aka.ms/sentinel-DigitalShadows-functionapp)"
         },
         {
             "title": "3. Configure the Function App",
