+++ Solutions/Digital Shadows/Data Connectors/host.json
