+++ Solutions/Digital Shadows/Data Connectors/proxies.json
