+++ Solutions/Digital Shadows/Data Connectors/readme.md
