+++ Solutions/Digital Shadows/Data Connectors/requirements.txt
