+++ Solutions/Digital Shadows/Data/Solution_DigitalShadowsSearchlight.json
@@ -14,7 +14,7 @@
 		"Playbooks/DigitalShadowsPlaybooktoUpdateIncidentStatus/DigitalShadowsPlaybook-UpdateIncidentStatus.json"
 	],
     "Data Connectors": [
-		"Data Connectors/Digital Shadows/DigitalShadowsSearchlight_API_functionApp.json"
+		"Data Connectors/DigitalShadowsSearchlight_API_functionApp.json"
 	],
     "BasePath": "C:\\GitHub\\Azure-Sentinel\\Solutions\\Digital Shadows",
     "Version": "3.0.0",
