+++ Solutions/Digital Shadows/ReleaseNotes.md
@@ -1,4 +1,5 @@
 | **Version** | **Date Modified (DD-MM-YYYY)** | **Change History**                                                 |
 |-------------|--------------------------------|--------------------------------------------------------------------|
-| 3.0.0       | 30-11-2023                     |  Added new Entity Mapping to **Analytic Rules**                    |
+| 3.0.0       | 04-06-2026                     |  Updated EventReportUrl construction in both **Analytic Rules**                    |
+|       | 30-11-2023                     |  Added new Entity Mapping to **Analytic Rules**                    |
